refactor(spend): split sqlite.rs into schema/recorder/pricing/query modules
Break the 1063-line sqlite.rs into a sqlite/ module: schema (open +
migrations + seed), recorder (async writer + SpendRecorder impl), pricing
(cost computation + table editor helpers), and query (filters +
aggregates). The root keeps the module docs, the now_unix helper, and the
test suite, and re-exports the public API flat so callers are unchanged.
